starting jupyter notebook

+++ ch02/fetch_data.py
@@ -10,12 +10,15 @@ HOUSING_URL = DOWNLOAD_ROOT + "datasets/housing/housing.tgz"
 
 
 def fetch_housing_data(housing_url=HOUSING_URL, housing_path=HOUSING_PATH):
-    os.makedirs(housing_path, exist_ok=True)
     tgz_path = os.path.join(housing_path, "housing.tgz")
-    urllib.request.urlretrieve(housing_url, tgz_path)
-    housing_tgz = tarfile.open(tgz_path)
-    housing_tgz.extractall(path=housing_path)
-    housing_tgz.close()
+    if os.path.exists(tgz_path):
+        print('file has already fetched')
+    else:
+        os.makedirs(housing_path, exist_ok=True)
+        urllib.request.urlretrieve(housing_url, tgz_path)
+        housing_tgz = tarfile.open(tgz_path)
+        housing_tgz.extractall(path=housing_path)
+        housing_tgz.close()
 
 
 def load_housing_data(housing_path=HOUSING_PATH):
